+++ src/hgraph/nodes/_numpy.py
@@ -1,10 +1,16 @@
 from dataclasses import dataclass
 from datetime import datetime
 
-from hgraph import compute_node, TS, TSB, SCALAR, TimeSeriesSchema, Array, SIZE, STATE, AUTO_RESOLVE, MIN_DT, NUMBER
 import numpy as np
 
+from hgraph import compute_node, TS, TSB, SCALAR, TimeSeriesSchema, Array, SIZE, STATE, AUTO_RESOLVE, MIN_DT, NUMBER, \
+    COMPOUND_SCALAR, Frame, Size
 
+
+__all__ = ("NpRollingWindowResult", "NpRollingWindowState", "np_rolling_window", "np_quantile", "np_std")
+
+
+@dataclass
 class NpRollingWindowResult(TimeSeriesSchema):
     buffer: TS[Array[SCALAR, SIZE]]
     index: TS[Array[datetime, SIZE]]
@@ -35,7 +41,7 @@ def np_rolling_window(ts: TS[SCALAR], period: SIZE, min_window_period: int = Non
         _state.start = start
         length = capacity
     _state.length = length
-    pos = (start+length-1) % capacity
+    pos = (start + length - 1) % capacity
     buffer[pos] = ts.value
     index[pos] = ts.last_modified_time
     if length == capacity or (min_window_period is not None and length >= min_window_period):
@@ -57,6 +63,35 @@ def np_rolling_window_start(_sz: type[SIZE], _scalar: type[SCALAR], _state: STAT
 
 
 @compute_node
-def np_quantile(ts: TS[Array[NUMBER]], q: float, method: str = 'linear') -> TS[float]:
+def np_quantile(ts: TS[Array[NUMBER, SIZE]], q: float, method: str = 'linear') -> TS[float]:
     """The np.quantile function, limited to a single axis"""
     return np.quantile(ts.value, q, method=method)
+
+
+@compute_node
+def np_std(ts: TS[Array[NUMBER, SIZE]]) -> TS[float]:
+    """Return the standard deviation of the values of the ts"""
+    return np.std(ts.value)
+
+
+def _compute_size(compound_type: COMPOUND_SCALAR) -> type:
+    schema = compound_type.__meta_data_schema__
+    values = iter(schema.values())
+    v = next(values)
+    if not all(v != v_ for v_ in values):
+        raise ValueError('Not all values of the frame are the same')
+    return Size[len(schema.values())]
+
+
+def _compute_data_tp(compound_type: COMPOUND_SCALAR) -> type:
+    schema = compound_type.__meta_data_schema__
+    values = iter(schema.values())
+    v = next(values)
+    return v.py_type
+
+
+@compute_node(resolvers={SIZE: lambda mapping, scalars: _compute_size(mapping[COMPOUND_SCALAR]),
+                         NUMBER: lambda mapping, scalars: _compute_data_tp(mapping[COMPOUND_SCALAR])})
+def frame_to_1d_array(frame: TS[Frame[COMPOUND_SCALAR]], _sz: type[SIZE] = Size, _tp: SCALAR = float) \
+        -> TS[Array[SCALAR, SIZE]]:
+    ...